requirements.txt
matplotlib==3.3.1
mysql-connector-python==8.0.21
tk==0.1.0